feat(settings-modals): split Privacy/Security tabs; section icons; mobile pills nav
- Split the User Settings "Privacy/Security" tab into Privacy (toggles) and
  Security (devices, key export, mobile import, data backup). New `eye` icon
  for Privacy, `shield` stays on Security. Authorized Devices section gets a
  short tooltip clarifying that Remove stops new messages but doesn't log the
  device out or wipe data.
- Reorder Privacy switches by importance: sync, signing, public profile up
  top, then a dashed separator, then receipts/typing/YouTube.
- Add the section icon next to each section title in both Space Settings and
  User Settings (mirrors the sidebar icons).
- Space Settings sidebar header now shows the space icon + space name in its
  natural casing instead of "SPACE SETTINGS"; User Settings reads
  "Account Settings".
- Replace the 2-column mobile menu with a horizontally-scrollable pill nav.
  Active pill gets an accent border + tinted background. Scroll-snap +
  scroll-into-view hook (skips first render so the natural scroll-at-zero
  position is preserved on mount). Trailing spacer keeps the last pill clear
  of the right-edge fade.
- Misc: small breathing room above the bio separator in User Settings
  General.
